springboot sample code with Kotlin API
This source code is written in Kotlin. It is a bout developing a webservice using SpringBoot. This code was introduced or developed as a first attmpt to learn Springboot autonomously "by myself". furthermore, it was a the first chance or to say first attempt to learn Kotlin API which is a great and powerful langauge

Please note that:
For the file "Controller1.kt" is to be palced in a folder called "controllers"

Looking forward hear and to receive your commit.